dev правки по замечаниям
