+++ server/internal/app/db/db/db.go
@@ -2,17 +2,12 @@ package db
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
-	"fmt"
 	"time"
 
-	"github.com/Alandres998/go-keeper/models"
 	"github.com/Alandres998/go-keeper/server/internal/app/db/storage"
 	"github.com/jmoiron/sqlx"
 	_ "github.com/lib/pq"
 	"go.uber.org/zap"
-	"golang.org/x/crypto/bcrypt"
 )
 
 // DBStorage для работы с базой данных.
@@ -95,171 +90,4 @@ func (s *DBStorage) execContext(ctx context.Context, tx *sqlx.Tx) sqlx.ExtContex
 		return tx
 	}
 	return s.db
-}
-
-// CreateUser создает нового пользователя в базе данных
-func (s *DBStorage) CreateUser(ctx context.Context, tx *sqlx.Tx, user *models.User) (*models.User, error) {
-	query := `
-		INSERT INTO users (login, password, email, created_at, updated_at)
-		VALUES ($1, $2, $3, $4, $5)
-		RETURNING id, created_at, updated_at
-	`
-	now := time.Now()
-	var err error
-
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(user.Password), bcrypt.DefaultCost)
-	if err != nil {
-		return nil, fmt.Errorf("ошибка хеширования пароля: %v", err)
-	}
-	user.Password = string(hashedPassword)
-	// Используем транзакцию, если она передана
-	if tx != nil {
-		err = tx.QueryRowContext(ctx, query, user.Login, user.Password, user.Email, now, now).Scan(&user.ID, &user.CreatedAt, &user.UpdatedAt)
-	} else {
-		err = s.db.QueryRowContext(ctx, query, user.Login, user.Password, user.Email, now, now).Scan(&user.ID, &user.CreatedAt, &user.UpdatedAt)
-	}
-
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("не удалось создать пользователя, никаких строк не вернулось: %v", err)
-		}
-		return nil, fmt.Errorf("не смог создать пользователя: %v", err)
-	}
-
-	return user, nil
-}
-
-// GetUserByLogin получает пользователя по логину
-func (s *DBStorage) GetUserByLogin(ctx context.Context, tx *sqlx.Tx, login string) (*models.User, error) {
-	query := `
-		SELECT id, login, password, email, created_at, updated_at
-		FROM users
-		WHERE login = $1
-	`
-	var user models.User
-	var err error
-
-	// Используем транзакцию, если она передана
-	if tx != nil {
-		err = tx.GetContext(ctx, &user, query, login)
-	} else {
-		err = s.db.GetContext(ctx, &user, query, login)
-	}
-
-	if err != nil {
-		return nil, fmt.Errorf("не корректный логин или пароль: %v", err)
-	}
-
-	return &user, nil
-}
-
-// AddSyncHistory добавляет запись о синхронизации данных
-func (s *DBStorage) AddSyncHistory(ctx context.Context, tx *sqlx.Tx, history *models.SyncHistory) (*models.SyncHistory, error) {
-	query := `
-		INSERT INTO sync_history (user_id, sync_timestamp, data_version, operation_type, meta_info)
-		VALUES ($1, $2, $3, $4, $5)
-		RETURNING id, sync_timestamp
-	`
-
-	now := time.Now()
-
-	metaInfoJSON, err := json.Marshal(history.MetaInfo)
-	if err != nil {
-		return nil, fmt.Errorf("ошибка сериализации MetaInfo: %v", err)
-	}
-
-	// Используем транзакцию, если она передана
-	if tx != nil {
-		err = tx.QueryRowContext(ctx, query, history.UserID, now, history.DataVersion, history.OperationType, string(metaInfoJSON)).Scan(&history.ID, &history.SyncTimestamp)
-	} else {
-		err = s.db.QueryRowContext(ctx, query, history.UserID, now, history.DataVersion, history.OperationType, string(metaInfoJSON)).Scan(&history.ID, &history.SyncTimestamp)
-	}
-
-	if err != nil {
-		return nil, fmt.Errorf("не смог загрузить историю: %v", err)
-	}
-
-	history.SyncTimestamp = now
-	return history, nil
-}
-
-// CountUserLogins возвращает количество попыток входа для пользователя
-func (s *DBStorage) CountUserLogins(ctx context.Context, tx *sqlx.Tx, userID int) (int, error) {
-	var count int
-	query := `SELECT COUNT(*) FROM sync_history WHERE user_id = $1 AND operation_type = 'login'`
-
-	// Используем транзакцию, если она передана
-	if tx != nil {
-		err := tx.QueryRowContext(ctx, query, userID).Scan(&count)
-		if err != nil {
-			return 0, err
-		}
-	} else {
-		err := s.db.QueryRowContext(ctx, query, userID).Scan(&count)
-		if err != nil {
-			return 0, err
-		}
-	}
-
-	return count, nil
-}
-
-// GetPrivateDataByUserID возвращает данные пользователя по userID
-func (s *DBStorage) GetPrivateDataByUserID(ctx context.Context, userID int) (*models.PrivateData, error) {
-	query := `
-		SELECT id, user_id, card_number, text_data, binary_data, meta_info
-		FROM private_data
-		WHERE user_id = $1
-		LIMIT 1
-	`
-
-	var data models.PrivateData
-	row := s.db.QueryRowContext(ctx, query, userID)
-
-	// Сканируем полученные данные в структуру
-	err := row.Scan(&data.ID, &data.UserID, &data.CardNumber, &data.TextData, &data.BinaryData, &data.MetaInfo)
-	if err == sql.ErrNoRows {
-		// Если данных нет, возвращаем nil и ошибку
-		return nil, nil
-	} else if err != nil {
-		// Обрабатываем другие ошибки
-		return nil, err
-	}
-
-	return &data, nil
-}
-
-// InsertOrUpdatePrivateData добавляет или обновляет запись в таблице private_data.
-func (s *DBStorage) InsertOrUpdatePrivateData(ctx context.Context, tx *sqlx.Tx, data *models.PrivateData) (*models.PrivateData, error) {
-	query := `
-		INSERT INTO private_data (user_id, text_data, binary_data, card_number, created_at, updated_at, meta_info)
-		VALUES ($1, $2, $3, $4, NOW(), NOW(), $5)
-		ON CONFLICT (user_id)
-		DO UPDATE SET
-			text_data = EXCLUDED.text_data,
-			binary_data = EXCLUDED.binary_data,
-			card_number = EXCLUDED.card_number,
-			meta_info = EXCLUDED.meta_info,
-			updated_at = NOW()
-		RETURNING id, created_at, updated_at
-	`
-
-	metaInfoJSON, err := json.Marshal(data.MetaInfo)
-	if err != nil {
-		return nil, fmt.Errorf("ошибка сериализации meta_info: %v", err)
-	}
-
-	var row *sql.Row
-	if tx != nil {
-		row = tx.QueryRowContext(ctx, query, data.UserID, data.TextData, data.BinaryData, data.CardNumber, metaInfoJSON)
-	} else {
-		row = s.db.QueryRowContext(ctx, query, data.UserID, data.TextData, data.BinaryData, data.CardNumber, metaInfoJSON)
-	}
-
-	err = row.Scan(&data.ID, &data.CreatedAt, &data.UpdatedAt)
-	if err != nil {
-		return nil, fmt.Errorf("ошибка выполнения запроса: %v", err)
-	}
-
-	return data, nil
 }
