+++ server/internal/app/routing/v1/options.go
@@ -10,6 +10,7 @@ type OptionsServiceServer struct {
 	options.UnimplementedOptionsServiceServer
 }
 
+// Метод для проверки клиентом доступности сервера
 func (s *OptionsServiceServer) Ping(ctx context.Context, req *options.PingRequest) (*options.PingResponse, error) {
 	return &options.PingResponse{Message: "Pong"}, nil
 }
